Conflicts:
	annotations/com/intellij/psi/search/searches/annotations.xml
	build.xml
	idea/src/org/jetbrains/jet/plugin/formatter/JetFormattingModelBuilder.java
	idea/testData/hierarchy/calls/callers/kotlinProperty/KotlinProperty_verification.xml
	idea/testData/hierarchy/calls/callers/kotlinProperty/main0.kt
	idea/testData/hierarchy/calls/callers/kotlinProperty/main1.java
	idea/tests/org/jetbrains/jet/checkers/KotlinAndJavaCheckerTest.java
	idea/tests/org/jetbrains/jet/plugin/parameterInfo/MockParameterInfoUIContext.java
	update_dependencies.xml

+++ idea/builtinsSrc/jet/Numbers.kt
@@ -387,8 +387,8 @@ public class Short private () : Number, Comparable<Short> {
 
   public fun inc() : Short
   public fun dec() : Short
-  public fun plus() : Short
-  public fun minus() : Short
+  public fun plus() : Int
+  public fun minus() : Int
 
   public override fun toDouble() : Double
   public override fun toFloat() : Float
@@ -461,8 +461,8 @@ public class Byte private () : Number, Comparable<Byte> {
 
   public fun inc() : Byte
   public fun dec() : Byte
-  public fun plus() : Byte
-  public fun minus() : Byte
+  public fun plus() : Int
+  public fun minus() : Int
 
   public override fun toDouble() : Double
   public override fun toFloat() : Float
